refactor: use `gridSelection` state by column itself
